refactor(console): remove legacy notification settings page
