+++ gravitee-apim-console-webui/src/components/notification-settings/notification-settings-details/notification-settings-details.component.html
