+++ gravitee-apim-console-webui/src/components/notification-settings/notifications-settings-add-dialog/notification-settings-add-dialog.component.ts
@@ -17,9 +17,9 @@ import { Component, Inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 
-import { Notifier } from '../../../../entities/notification/notifier';
-import { NewNotificationSettings } from '../../../../entities/notification/newNotificationSettings';
-import { NotificationSettings } from '../../../../entities/notification/notificationSettings';
+import { Notifier } from '../../../entities/notification/notifier';
+import { NewNotificationSettings } from '../../../entities/notification/newNotificationSettings';
+import { NotificationSettings } from '../../../entities/notification/notificationSettings';
 
 type Reference = {
   referenceType: 'API' | 'APPLICATION' | 'PORTAL';
