+++ gravitee-apim-console-webui/src/management/api/notification-settings/notification-settings-list/api-notification-settings-list.component.ts
@@ -19,7 +19,7 @@ import { Subject } from 'rxjs';
 import { ActivatedRoute } from '@angular/router';
 
 import { ApiNotificationSettingsService } from '../../../../services-ngx/api-notification-settings.service';
-import { NotificationSettingsListServices } from '../../../../components/notifications/notification-settings/notification-settings-list.component';
+import { NotificationSettingsListServices } from '../../../../components/notification-settings/notification-settings-list.component';
 
 @Component({
   selector: 'api-notification-settings-list',
